requirements.txt
torch==0.4.1
torchvision==0.2.1
albumentations
pretrainedmodels
tqdm
opencv-python
pillow-simd
numpy
sklearn
pandas
matplotlib
seaborn
iterative-stratification
pystruct
cvxopt